% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Water.R
\name{Detect_water}
\alias{Detect_water}
\title{Title}
\usage{
Detect_water(x)
}
\arguments{
\item{x}{Must be a .tif file
Checks dimensions
Diverts by if into RGB and RGBA
Performs K-means clustering
Segments data "Fig" raster into three matrices by color channel
Replaces matrix values with the K mean for each color
Reshapes matricies so they match the original raster
Converts back to Raster}
}
\value{
Plot of color segmented raster with green = blue
}
\description{
Title
}
