% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind 2 factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{Factor a}

\item{b}{Factor b}
}
\value{
A new factor that binds a and b
}
\description{
Create a new factor from 2 existing factors where there are
a union of the levels of existing factors
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
